refactor: 로딩 아이콘을 공용 컴포넌트로 변경
