+++ apps/mac/src/renderer/src/shared/ui/clash-loading-icon/ClashLoadingIcon.tsx
