+++ apps/mac/src/renderer/src/shared/ui/index.ts
@@ -28,3 +28,4 @@ export { DefaultProfileIcon, defaultProfileImageDark, ProfileAvatar } from "./pr
 export type { ProfileAvatarProps } from "./profile-avatar";
 export { NameTag } from "./name-tag";
 export type { NameTagProps, NameTagSize, NameTagTextAlign, NameTagWidth } from "./name-tag";
+export { ClashLoadingIcon } from "./clash-loading-icon/ClashLoadingIcon";
